Refactored searchFilter with two new folders for holding components related to currentTags and searchSelect. Also created new tagTemplate file in order for remove tag icons to work individually as intended
